Viết tầng service cho dự án

+++ ShopExample.Data/Infrastructure/IRepository.cs
@@ -18,13 +18,13 @@ namespace ShopExample.Data.Infrastructure
         // Marks an entity to be removed
         T Delete(T entity);
 
-        T Delete(int id);
+        T Delete(long id);
 
         //Delete multi records
         void DeleteMulti(Expression<Func<T, bool>> where);
 
         // Get an entity by int id
-        T GetSingleById(int id);
+        T GetSingleById(long id);
 
         T GetSingleByCondition(Expression<Func<T, bool>> expression, string[] includes = null);
